+++ migrations/postgres/001_initial.sql
@@ -5,7 +5,7 @@
 CREATE TABLE proven_txs (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "provenTxId" SERIAL PRIMARY KEY,
+    "provenTxId" BIGSERIAL PRIMARY KEY,
     txid VARCHAR(64) NOT NULL UNIQUE,
     height INTEGER NOT NULL,
     "index" INTEGER NOT NULL,
@@ -19,8 +19,8 @@ CREATE INDEX idx_proven_txs_blockHash ON proven_txs("blockHash");
 CREATE TABLE proven_tx_reqs (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "provenTxReqId" SERIAL PRIMARY KEY,
-    "provenTxId" INTEGER REFERENCES proven_txs("provenTxId"),
+    "provenTxReqId" BIGSERIAL PRIMARY KEY,
+    "provenTxId" BIGINT REFERENCES proven_txs("provenTxId"),
     status VARCHAR(16) NOT NULL DEFAULT 'unknown',
     attempts INTEGER NOT NULL DEFAULT 0,
     notified BOOLEAN NOT NULL DEFAULT FALSE,
@@ -38,7 +38,7 @@ CREATE INDEX idx_proven_tx_reqs_txid ON proven_tx_reqs(txid);
 CREATE TABLE users (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "userId" SERIAL PRIMARY KEY,
+    "userId" BIGSERIAL PRIMARY KEY,
     "identityKey" VARCHAR(130) NOT NULL UNIQUE,
     "activeStorage" VARCHAR(130) NOT NULL
 );
@@ -46,8 +46,8 @@ CREATE TABLE users (
 CREATE TABLE certificates (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "certificateId" SERIAL PRIMARY KEY,
-    "userId" INTEGER NOT NULL REFERENCES users("userId"),
+    "certificateId" BIGSERIAL PRIMARY KEY,
+    "userId" BIGINT NOT NULL REFERENCES users("userId"),
     "serialNumber" VARCHAR(100) NOT NULL,
     type VARCHAR(100) NOT NULL,
     certifier VARCHAR(100) NOT NULL,
@@ -62,8 +62,8 @@ CREATE TABLE certificates (
 CREATE TABLE certificate_fields (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "userId" INTEGER NOT NULL REFERENCES users("userId"),
-    "certificateId" INTEGER NOT NULL REFERENCES certificates("certificateId"),
+    "userId" BIGINT NOT NULL REFERENCES users("userId"),
+    "certificateId" BIGINT NOT NULL REFERENCES certificates("certificateId"),
     "fieldName" VARCHAR(100) NOT NULL,
     "fieldValue" VARCHAR(255) NOT NULL,
     "masterKey" VARCHAR(255) NOT NULL DEFAULT '',
@@ -73,8 +73,8 @@ CREATE TABLE certificate_fields (
 CREATE TABLE output_baskets (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "basketId" SERIAL PRIMARY KEY,
-    "userId" INTEGER NOT NULL REFERENCES users("userId"),
+    "basketId" BIGSERIAL PRIMARY KEY,
+    "userId" BIGINT NOT NULL REFERENCES users("userId"),
     name VARCHAR(300) NOT NULL,
     "numberOfDesiredUTXOs" INTEGER NOT NULL DEFAULT 6,
     "minimumDesiredUTXOValue" INTEGER NOT NULL DEFAULT 10000,
@@ -85,9 +85,9 @@ CREATE TABLE output_baskets (
 CREATE TABLE transactions (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "transactionId" SERIAL PRIMARY KEY,
-    "userId" INTEGER NOT NULL REFERENCES users("userId"),
-    "provenTxId" INTEGER REFERENCES proven_txs("provenTxId"),
+    "transactionId" BIGSERIAL PRIMARY KEY,
+    "userId" BIGINT NOT NULL REFERENCES users("userId"),
+    "provenTxId" BIGINT REFERENCES proven_txs("provenTxId"),
     status VARCHAR(64) NOT NULL,
     reference VARCHAR(64) NOT NULL UNIQUE,
     "isOutgoing" BOOLEAN NOT NULL,
@@ -105,10 +105,10 @@ CREATE INDEX idx_transactions_txid ON transactions(txid);
 CREATE TABLE commissions (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "commissionId" SERIAL PRIMARY KEY,
-    "userId" INTEGER NOT NULL REFERENCES users("userId"),
-    "transactionId" INTEGER NOT NULL UNIQUE REFERENCES transactions("transactionId"),
-    satoshis INTEGER NOT NULL,
+    "commissionId" BIGSERIAL PRIMARY KEY,
+    "userId" BIGINT NOT NULL REFERENCES users("userId"),
+    "transactionId" BIGINT NOT NULL UNIQUE REFERENCES transactions("transactionId"),
+    satoshis BIGINT NOT NULL,
     "keyOffset" VARCHAR(130) NOT NULL,
     "isRedeemed" BOOLEAN NOT NULL DEFAULT FALSE,
     "lockingScript" BYTEA NOT NULL
@@ -118,10 +118,10 @@ CREATE INDEX idx_commissions_transactionId ON commissions("transactionId");
 CREATE TABLE outputs (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "outputId" SERIAL PRIMARY KEY,
-    "userId" INTEGER NOT NULL REFERENCES users("userId"),
-    "transactionId" INTEGER NOT NULL REFERENCES transactions("transactionId"),
-    "basketId" INTEGER REFERENCES output_baskets("basketId"),
+    "outputId" BIGSERIAL PRIMARY KEY,
+    "userId" BIGINT NOT NULL REFERENCES users("userId"),
+    "transactionId" BIGINT NOT NULL REFERENCES transactions("transactionId"),
+    "basketId" BIGINT REFERENCES output_baskets("basketId"),
     spendable BOOLEAN NOT NULL DEFAULT FALSE,
     change BOOLEAN NOT NULL DEFAULT FALSE,
     vout INTEGER NOT NULL,
@@ -135,7 +135,7 @@ CREATE TABLE outputs (
     "derivationPrefix" VARCHAR(200),
     "derivationSuffix" VARCHAR(200),
     "customInstructions" VARCHAR(2500),
-    "spentBy" INTEGER REFERENCES transactions("transactionId"),
+    "spentBy" BIGINT REFERENCES transactions("transactionId"),
     "sequenceNumber" INTEGER,
     "spendingDescription" VARCHAR(2048),
     "scriptLength" BIGINT,
@@ -152,8 +152,8 @@ CREATE INDEX idx_outputs_spentby ON outputs("spentBy");
 CREATE TABLE output_tags (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "outputTagId" SERIAL PRIMARY KEY,
-    "userId" INTEGER NOT NULL REFERENCES users("userId"),
+    "outputTagId" BIGSERIAL PRIMARY KEY,
+    "userId" BIGINT NOT NULL REFERENCES users("userId"),
     tag VARCHAR(150) NOT NULL,
     "isDeleted" BOOLEAN NOT NULL DEFAULT FALSE,
     UNIQUE(tag, "userId")
@@ -162,8 +162,8 @@ CREATE TABLE output_tags (
 CREATE TABLE output_tags_map (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "outputTagId" INTEGER NOT NULL REFERENCES output_tags("outputTagId"),
-    "outputId" INTEGER NOT NULL REFERENCES outputs("outputId"),
+    "outputTagId" BIGINT NOT NULL REFERENCES output_tags("outputTagId"),
+    "outputId" BIGINT NOT NULL REFERENCES outputs("outputId"),
     "isDeleted" BOOLEAN NOT NULL DEFAULT FALSE,
     UNIQUE("outputTagId", "outputId")
 );
@@ -173,8 +173,8 @@ CREATE INDEX idx_output_tags_map_output_deleted_tag ON output_tags_map("outputId
 CREATE TABLE tx_labels (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "txLabelId" SERIAL PRIMARY KEY,
-    "userId" INTEGER NOT NULL REFERENCES users("userId"),
+    "txLabelId" BIGSERIAL PRIMARY KEY,
+    "userId" BIGINT NOT NULL REFERENCES users("userId"),
     label VARCHAR(300) NOT NULL,
     "isDeleted" BOOLEAN NOT NULL DEFAULT FALSE,
     UNIQUE(label, "userId")
@@ -183,8 +183,8 @@ CREATE TABLE tx_labels (
 CREATE TABLE tx_labels_map (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "txLabelId" INTEGER NOT NULL REFERENCES tx_labels("txLabelId"),
-    "transactionId" INTEGER NOT NULL REFERENCES transactions("transactionId"),
+    "txLabelId" BIGINT NOT NULL REFERENCES tx_labels("txLabelId"),
+    "transactionId" BIGINT NOT NULL REFERENCES transactions("transactionId"),
     "isDeleted" BOOLEAN NOT NULL DEFAULT FALSE,
     UNIQUE("txLabelId", "transactionId")
 );
@@ -194,7 +194,7 @@ CREATE INDEX idx_tx_labels_map_tx_deleted ON tx_labels_map("transactionId", "isD
 CREATE TABLE monitor_events (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     event VARCHAR(64) NOT NULL,
     details TEXT
 );
@@ -213,8 +213,8 @@ CREATE TABLE settings (
 CREATE TABLE sync_states (
     created_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP(3) NOT NULL DEFAULT NOW(),
-    "syncStateId" SERIAL PRIMARY KEY,
-    "userId" INTEGER NOT NULL REFERENCES users("userId"),
+    "syncStateId" BIGSERIAL PRIMARY KEY,
+    "userId" BIGINT NOT NULL REFERENCES users("userId"),
     "storageIdentityKey" VARCHAR(130) NOT NULL DEFAULT '',
     "storageName" TEXT NOT NULL,
     status TEXT NOT NULL DEFAULT 'unknown',
